+++ src/psolver2.hpp
@@ -31,7 +31,7 @@ using namespace FETools;
 class parallelSolver2
 {
 private:
-  static constexpr unsigned int DIM = 2;
+  static constexpr unsigned short DIM = 2;
   //first of all, the mesh that will be saved as a member of solver,
   Mesh<DIM> _mesh;
   // a fe class solver object, to deal with the single elements of the mesh member
